# docs/swagger.yaml
basePath: /
definitions:
...
  modelscar.Car_Model:
    properties:
      car_number:
        type: string
      duration:
        type: integer
      end_time:
        type: string
      id:
        type: integer
      image_url:
        type: string
      park_no:
        type: string
      reason:
        type: string
      start_time:
        type: string
      status:
        type: string
      total_payment:
        type: number
      user_id:
        type: string
    type: object
  modelscar.CarUpdate:
    properties:
      reason:
        type: string
    type: object
...
  operator.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  operator.GetCarsResponse:
    properties:
      cars:
        items:
          $ref: '#/definitions/modelscar.Car_Model'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total_count:
        type: integer
    type: object
...
host: 192.168.100.7:3000
info:
  contact: {}
  title: Airline REST API
paths:
...
  /api/v1/camera/updatecar/{plate}:
    put:
      consumes:
      - application/json
      description: Updates a car's status and calculates payment and duration based
        on start and end times.
      parameters:
      - description: Car plate number
        in: path
        name: plate
        required: true
        type: string
      - description: Car details to update
        in: body
        name: car
        required: true
        schema:
          $ref: '#/definitions/modelscar.CarUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Updated car details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Car already exited or invalid request
          schema:
            $ref: '#/definitions/operator.ErrorResponse'
        "404":
          description: Car not found
          schema:
            $ref: '#/definitions/operator.ErrorResponse'
        "500":
          description: Error parsing time
          schema:
            $ref: '#/definitions/operator.ErrorResponse'
      summary: Update a car by plate number
      tags:
      - cars
...
  /api/v1/getallcars:
    get:
      consumes:
      - application/json
      description: Get list of cars with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 5
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/operator.GetCarsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/operator.ErrorResponse'
      summary: Get list of cars
      tags:
      - cars
  /api/v1/getcar/{id}:
    get:
      consumes:
      - application/json
      description: Get a car by ID
      parameters:
      - description: Car ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/modelscar.Car_Model'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/operator.ErrorResponse'
      summary: Get a car by ID
      tags:
      - cars
  /api/v1/searchcar:
    get:
      consumes:
      - application/json
      description: Search for a car by plate number, parking number, and other optional
        filters
      parameters:
      - description: Car plate number
        in: query
        name: car_number
        type: string
      - description: Enter time (YYYY-MM-DD)
        in: query
        name: enter_time
        type: string
      - description: End time (YYYY-MM-DD)
        in: query
        name: end_time
        type: string
      - description: Parking spot number
        in: query
        name: parkno
        type: string
      - description: Car status (Inside, Exited)
        in: query
        name: status
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 5
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/operator.GetCarsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/operator.ErrorResponse'
      summary: Search for a car by plate number and optional filters
      tags:
      - cars
...
swagger: "2.0"